% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hmm.R
\name{build_hmm}
\alias{build_hmm}
\title{Build a model to assign haplotypes to each gamete}
\usage{
build_hmm(num_snps, sequencing_error, avg_recomb)
}
\arguments{
\item{num_snps}{number of SNPs, found previously from taking the number of rows in the input data. Used to set the denominator for transition probability}

\item{sequencing_error}{User-input for expected error in sequencing (default = 0.005)}

\item{avg_recomb}{User-input for expected average recombination spots per chromosome (default=1). Used to set the numerator for transition probability}
}
\value{
hmm The hidden Markov model with transition and emission probabilities set for use.
}
\description{
This function builds a hidden Markov model that considers sequencing error. Uses the
HMM packages.
}
