% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_window_size.R
\name{calculate_window_size}
\alias{calculate_window_size}
\title{A function to automatically calculate a preferred window size for phasing}
\usage{
calculate_window_size(ngam, cov, nsnp, ger, avgr)
}
\arguments{
\item{ngam}{an integer; number of gametes in the dataset; this should be the number of columns in the input dataframe (excluding the positions column)}

\item{cov}{a numeric; sequencing depth of coverage applicable to the data}

\item{nsnp}{an integer; number of hetSNPs in the dataset; this should be the number of rows in the input dataframe (after filtering to include hetSNPs only)}

\item{ger}{a numeric; genotyping error rate applicable to data}

\item{avgr}{a numeric; average recombination rate}
}
\value{
window_size an integer; the preferred window size for accurate phasing given the data characteristics
}
\description{
This function given the number of gametes (\code{gametes}), coverage (\code{coverage}),
genotyping error rate (\code{se}), and average recombination rate (\code{avgr}),
as well as the number of hetSNPs in the data,
is used to automatically calculate a preferred window size for phasing.
This is accomplished by using the predict function and a pre-trained beta regression model \code{betaregmodel_20220718}.
The user can call this function themselves outside of rhapsodi's main steps, or the user can specify they would rather rhapsodi
calculate a window size based on the input data characteristics.
}
