% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_pos.R
\name{cbind_pos}
\alias{cbind_pos}
\title{This function adds an index and a pos column to dataframes}
\usage{
cbind_pos(input_df, position_vec)
}
\arguments{
\item{input_df}{the dataframe which the user wants to append index and pos columns to}

\item{position_vec}{the vector of positions which will be appended to the dataframe}
}
\value{
output_df the dataframe with two new columns at the beginning, index and pos, as well as the original columns
}
\description{
This function adds an index and a pos column to dataframes
}
