% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_meiotic_recombination.R
\name{discover_meiotic_recombination}
\alias{discover_meiotic_recombination}
\title{A function to drive and report meiotic recombination breakpoint finding}
\usage{
discover_meiotic_recombination(
  original_gamete_data,
  complete_haplotypes,
  filled_gamete_data_list,
  positions,
  smooth_crossovers = TRUE,
  smooth_imputed_genotypes = FALSE,
  sampleName = "sampleT",
  chrom = "chrT",
  threads = 2
)
}
\arguments{
\item{original_gamete_data}{original sparse gamete data matrix}

\item{complete_haplotypes}{dataframe of phased diploid donor genotypes in two columns, each column corresponding with a haplotype from the donor}

\item{filled_gamete_data_list}{the output list from \code{impute_gamete_genotypes} which contains each gamete data matrix with haplotype info from the HMM and \code{fill_NA} functions}

\item{positions}{the genomic positions corresponding to SNP indices}

\item{smooth_crossovers}{boolean, default is TRUE, whether to use smoothed data for recombination finding. If \code{TRUE}, doesn't replace with original reads}

\item{smooth_imputed_genotypes}{boolean, default is FALSE, whether to use smoothed data for ending genotypes. If \code{TRUE}, doesn't replace with original reads}

\item{sampleName}{sample name of sample given to rhapsodi, default is "sampleT"}

\item{chrom}{chromosome of sample given to rhapsodi, default is "chromT"}

\item{threads}{an integer, default = 2, the number of cores to use when we use mclapply or the like}
}
\value{
out a dataframe recomb_breaks, specifying the predicted recombination breakpoints for each gamete
}
\description{
This function takes as input two booleans controlling whether the gamete haplotypes and genotypes remain smoothed before crossover discovery (directly from HMM)
or unsmoothed (replacing inferred HMM state with the original reads if they disagree)
Then the function runs recombination finding
It offers the option to avoid oversmoothing by superimposing initial haplotype assignments over each gamete. For example, if an
allele assigned to h1 was changed by the model to h2, the early functions can fill the NAs to h2, but then un-smoothing
will replace the singular h1 at the correct allele. This singular h1 could be an example of gene conversion or non-crossover.
}
