% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gametes.R
\name{fill_gametes}
\alias{fill_gametes}
\title{A function to assign the haplotypes of each allele on every gamete}
\usage{
fill_gametes(
  dt,
  complete_haplotypes,
  sequencing_error = 0.005,
  avg_recomb = 1,
  threads = 2
)
}
\arguments{
\item{dt}{matrix of gametes}

\item{complete_haplotypes}{Inferred parental haplotypes}

\item{sequencing_error}{User-input for expected error in sequencing (default = 0.005)}

\item{avg_recomb}{User-input for average recombination rate that can be expected for a chromosome (default=1)}

\item{threads}{User-input value for calling \code{pbmclapply} or \code{mclapply} (default = 2)}
}
\value{
filled_gametes a tibble resulting from the HMM and fill_NAs function which has the imputed donor haplotypes for each gamete
}
\description{
This function builds and applies a hidden Markov model to categorize each allele on each gamete.
It then fills the positions missing data with the nearest haplotype assignment.
}
