% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_problem_genome.R
\name{filter_problem_genome}
\alias{filter_problem_genome}
\title{A function to filter user input data}
\usage{
filter_problem_genome(
  input_data_positions,
  chrom,
  exclude_genomic_positions,
  include_genomic_positions
)
}
\arguments{
\item{input_data_positions}{Bed-style file (with chr, start, end columns) showing genomic positions in the raw data}

\item{chrom}{Number of chromosome (can iterate through e.g., by calling in lapply)}

\item{exclude_genomic_positions}{Bed-style file (with chr, start, end columns) showing genomic positions that should be excluded (e.g., are in a blacklisted region of the genome)}

\item{include_genomic_positions}{Bed-style file (with chr, start, end columns) showing genomic positions that should be included (e.g., are in an accessible region of the genome)}
}
\value{
pos_not_excluded Genomic positions that should be included
}
\description{
This function takes in a datatable of user input data, with genomic positions, and either a file of genomic positions that should be included or a file of genomic positions that should be
excluded (or both!). It returns list of approved genomic positions. The user can then filter their raw data to only include regions that are within these approved ranges.
}
