% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_recomb_spots.R
\name{find_recomb_spots}
\alias{find_recomb_spots}
\title{A function to find recombination breakpoint regions for a single gamete}
\usage{
find_recomb_spots(input_gamete_data, x, identities, genomic_positions)
}
\arguments{
\item{input_gamete_data}{(tibble) SNPs correspond to the rows and gametes correspond to the columns}

\item{x}{which gamete number to search}

\item{identities}{gamete identities vector}

\item{genomic_positions}{Genomic SNP positions}
}
\value{
recomb_spots tibble of the recombination spots for gamete x
}
\description{
This function finds recombination breakpoint regions for a single gamete by finding
where adjacent non-NA haplotypes switch. Most likely large regions will be returned
rather than 2bp windows because the sparsity of the original data constrains how
many NAs will remain after filling and therefore we can only say that the true
exchange point occurs somewhere between these two non-matching haplotypes
}
