% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_donor_haplotypes.R
\name{impute_donor_haplotypes}
\alias{impute_donor_haplotypes}
\title{A function to impute haplotypes from diploid donor}
\usage{
impute_donor_haplotypes(
  dt,
  positions,
  window_length = 3000,
  overlap_denom = 2,
  threads = 2,
  mcstop = TRUE,
  stringent_stitch = TRUE,
  stitch_new_min = 0.5
)
}
\arguments{
\item{dt}{Matrix of gamete alleles}

\item{positions}{vector of SNP positions}

\item{window_length}{Size of window (default=3000)}

\item{overlap_denom}{User-input value for denominator in calculation of overlap (default = 2)}

\item{threads}{User-input value for calling \code{pbmclapply} (default = 2)}

\item{mcstop}{User-input value for continuing phasing even if mean_concordance isn't within advised bounds (default = TRUE)}

\item{stringent_stitch}{Boolean, whether or not to use the original bounds (0.1, 0.9 cutoff)}

\item{stitch_new_min}{Numeric/float, if \code{stringent_stitch} == FALSE, the new minimum value to use as a cutoff. If concordance is below this value, we assume the windows are opposite haplotypes, if greater than or equal, we assume the same haplotypes}
}
\value{
complete_haplotypes phased haplotypes as a tibble with column names index, pos (for SNP positions), h1 (haplotype 1), & h2 (haplotype 2)
}
\description{
This function infers the diploid donor haplotypes by first calling \code{split_with_overlap} to find the overlapping SNP segments,
then \code{reconstruct_hap} for majority voting within each overlapping segment, finally stitching together the overlapping segments to phase both haplotypes
}
