% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_gamete_genotypes.R
\name{impute_gamete_genotypes}
\alias{impute_gamete_genotypes}
\title{A function to drive the assignment and reporting of the genotypes of each allele on every gamete}
\usage{
impute_gamete_genotypes(
  original_gamete_data,
  complete_haplotypes,
  positions,
  sequencing_error = 0.005,
  avg_recomb = 1,
  smooth_imputed_genotypes = FALSE,
  threads = 2
)
}
\arguments{
\item{original_gamete_data}{original matrix of gametes}

\item{complete_haplotypes}{Inferred parental haplotypes}

\item{positions}{vector of SNP position}

\item{sequencing_error}{User-input for expected error in sequencing (default = 0.005)}

\item{avg_recomb}{User-input for average recombination rate that can be expected for a chromosome (default=1)s}

\item{smooth_imputed_genotypes}{a bool, default is FALSE, whether to use smoothed data for ending genotypes. If \code{TRUE}, doesn't replace with original reads, returning smoothed data only. If \code{FALSE}, will return both smoothed and unsmoothed}

\item{threads}{User-input value for calling \code{pbmclapply} or \code{mclapply} (default = 2)}
}
\value{
gamete_data a named list with two data frames (names filled_gametes and unsmoothed_gametes) resulting from the HMM, fill_NAs, and potentially the unsmooth function, returning the imputed donor haplotypes for each gamete. In the filled_gametes output, the dataframe represents the direct output. In the unsmoothed_gametes output, if \code{smooth_imputed_genotypes} is TRUE, this is NULL; if FALSE, original sequencing reads replace imputed genotypes if they disagree
}
\description{
This function builds and applies a hidden Markov model to categorize each allele on each gamete.
It then fills the positions missing data with the nearest haplotype assignment.
If the user asks for unsmoothed genotypes (i.e. replacing original sequencing reads where HMM imputation disagrees with these original reads) by setting \code{smooth_imputed_genotypes} to FALSE
then the unsmooth function is called to replace imputed genotypes with original sequencing reads
}
