% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_bits.R
\name{invert_bits}
\alias{invert_bits}
\title{A function to invert the values in a data frame or matrix}
\usage{
invert_bits(input_data)
}
\arguments{
\item{input_data}{gamete alleles}
}
\value{
input_data inverted from the actual input
}
\description{
This function replaces 0s with 1s and 1s with 0s in a dataframe or matrix
}
