% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_input.R
\name{nucleotide_input}
\alias{nucleotide_input}
\title{Function for rhapsodi to convert input of A/C/G/T to 0/1 (ref/alt)}
\usage{
nucleotide_input(info_and_gametes)
}
\arguments{
\item{info_and_gametes}{dataframe (no header) of genomic position, ref allele, alt allele, and read for each gamete at each position in nucleotide format}
}
\value{
input in a named list format with \code{gametes_het}, \code{positions}, \code{ref}, & \code{alt} where \code{gametes_het} is a data frame of the heterozygous sparse gamete data (in 0/1 format) and \code{positions} is a vector of the SNP genomic positions, \code{ref} is a vector of the reference allele for each SNP genomic position in A/C/G/T, and \code{alt} is a vector of the alternate allel for each SNP genomic position in A/C/G/T
}
\description{
A function to read in the sparse gamete sequencing data encoded in A/C/G/T tab-delimited format with the first column containing SNP positions in integer format, the second column containing the ref allele, and the third column containing the alt allele
}
\details{
This function is called when data is originally in nucleotide (A/C/G/T) format. The file is expected to be a tab-delimited file and to have the SNP genomic positions in the first column
We assume that a single file originates from a single sample and chromosome
Finally, we subset to include only heterozygous SNPs or those that have atleast one reference (0) and one alternate allele (1)
}
