% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gametes.R
\name{recode_gametes}
\alias{recode_gametes}
\title{A function to assign each allele to a haplotype or an NA if not enough information is known}
\usage{
recode_gametes(dt, complete_haplotypes)
}
\arguments{
\item{dt}{Input matrix of gamete reads}

\item{complete_haplotypes}{Inferred donor haplotypes}
}
\value{
dt matrix of gametes coded by haplotype at each position
}
\description{
This function reads along each gametes and replaces its read (0 or 1) with the corresponding
haplotype (h1 or h2) based on the allele in each donor haplotype; alternatively, the read
may be replaced with an NA if not enough information was known for phasing
}
