% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_hap.R
\name{reconstruct_haplotypes}
\alias{reconstruct_haplotypes}
\title{A function to reconstruct the donor haplotypes using data from gametes}
\usage{
reconstruct_haplotypes(input_dt, input_positions, window_indices)
}
\arguments{
\item{input_dt}{A dataframe of sparse gamete data coded with reference and alt alleles (i.e., 0, 1, or NA)}

\item{input_positions}{vector of SNP positions}

\item{window_indices}{A window of the gamete data (using segments generated from \code{split_with_overlap})}
}
\value{
inferred_output A tibble with the inferred haplotype sorted by position and window
}
\description{
This function takes gamete data and computes and clusters a distance matrix with a binary method. It
anticipates sparse data and replaces NAs with 0.5. It clusters the tree into two groups, i.e., haplotypes.
It then categorizes the gamete cells falling into each of the two groups. It reconstructs the original
haplotypes by majority vote after inverting the opposite haplotype.
}
