% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_gametes.R
\name{report_gametes}
\alias{report_gametes}
\title{A function to drive and report the unsmoothing (if desired) and recombination finding}
\usage{
report_gametes(
  smooth_crossovers,
  smooth_imputed_genotypes,
  complete_haplotypes,
  original_gamete_data,
  filled_gamete_data,
  positions,
  sampleName,
  chrom,
  threads = 2
)
}
\arguments{
\item{smooth_crossovers}{boolean whether to use smoothed data for recombination finding. If \code{TRUE}, doesn't replace with original reads}

\item{smooth_imputed_genotypes}{boolean whether to use smoothed data for ending genotypes. If \code{TRUE}, doesn't replace with original reads}

\item{complete_haplotypes}{dataframe of phased diploid donor genotypes in two columns, each column corresponding with a haplotype from the donor}

\item{original_gamete_data}{original sparse gamete data matrix}

\item{filled_gamete_data}{gamete data matrix from the HMM and \code{fill_NA} function}

\item{positions}{the genomic positions corresponding to SNP indices}

\item{sampleName}{sample name of sample given to rhapsodi}

\item{chrom}{chromosome of sample given to rhapsodi}

\item{threads}{an integer, default = 2, the number of cores to use when we use mclapply or the like}
}
\value{
out a named list which returns gamete_haps, or a data frame specifying from which donor haplotype each gamete position originates
gamete_genotypes, or a matrix specifying the genotype in (0's and 1's) for each gamete position
recomb_breaks, or a tibble specifying the predicted recombination breakpoints for each gamete
donor_haps, or phased haplotypes as a tibble with column names: index, pos (for SNP positions), h1 (haplotype 1), & h2 (haplotype 2)
}
\description{
This function takes as innput two booleans controlling wehther the reported genotypes are smoothed
or unsmoothed (replacing inferred HMM state with the original reads if they disagree)
and whether the data given to recombination finding is smoothed or unsmoothed. Then the function runs recombination finding
and eventually drives repporting/exporting the cool data
It offers the option to avoid oversmoothing by superimposing initial haplotype assignments over each gamete. For example, if an
allele assigned to h1 was changed by the model to h2, the early functions can fill the NAs to h2, but then un-smoothing
will replace the singular h1 at the correct allele. This singular h1 could be an example of gene conversion or non-crossover.
}
