% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhapsodi_autorun.R
\name{rhapsodi_autorun}
\alias{rhapsodi_autorun}
\title{This function can be used to run all steps of rhapsodi}
\usage{
rhapsodi_autorun(
  input_file,
  use_dt = FALSE,
  input_dt = NULL,
  acgt = FALSE,
  threads = 2,
  sampleName = "sampleT",
  chrom = "chrT",
  seqError_model = 0.005,
  avg_recomb_model = 1,
  window_length = 3000,
  overlap_denom = 2,
  mcstop = TRUE,
  stringent_stitch = TRUE,
  stitch_new_min = 0.5,
  smooth_imputed_genotypes = FALSE,
  smooth_crossovers = TRUE
)
}
\arguments{
\item{input_file}{a string; the path plus filename for the input sparse gamete genotype data in tabular form. Note the form is different depending on the value of \code{acgt}. Use NULL if \code{use_dt} is TRUE}

\item{use_dt}{a bool; default is FALSE, whether to input a pre-loaded data frame/table rather than using an input file}

\item{input_dt}{a data frame/table; only necessary if use_dt is TRUE. User-pre-loaded data frame/table. Note the format is different depending on the value of \code{acgt}}

\item{acgt}{a bool; default is FALSE; If TRUE, assumes that the data is not 0/1 encoded and is available in pre-loaded data frame format passed with \code{input_dt}}

\item{threads}{an integer; default is 2, number of threads to utilize when we use \code{mclapply}}

\item{sampleName}{a string; default is "sampleT", fill in with whatever the sample name is. We assume a single input file is from a single file}

\item{chrom}{a string; default is "chrT", fill in with whatever the chromosome is. We assume a single input file is from a single chromosome}

\item{seqError_model}{a numeric; default is 0.005, used in \code{build_hmm} within \code{fill_gametes}, the exppected error in sequencing}

\item{avg_recomb_model}{a numeric; default is 1, used in \code{build_hmm} within \code{fill_gametes}, the expected number of average recombination spots per chromosome}

\item{window_length}{an integer; default is 3000, used in \code{split_with_overlap} within \code{impute_donor_haplotypes}, the segment length to use in constructing overlapping windows for phasing}

\item{overlap_denom}{an integer; default is 2, used in \code{split_with_overlap} within \code{impute_donor_haplotypes}, User-input value for denominator in calculation of overlap, or the degree of overlap between segments}

\item{mcstop}{a bool; used in \code{stitch_haplotypes} within \code{impute_donor_haplotypes}, only considered if \code{stringent_stitch} is TRUE; default is TRUE; this parameter is used to determine whether phasing continues or exits if the mean concordance between two windows is between 0.1 and 0.9. If TRUE, rhapsodi exits. If FALSE, rhapsodi and phasing continues, asking which threshold the concordance is closer to and acting accordingly}

\item{stringent_stitch}{a bool; used in \code{stitch_haplotypes} within \code{impute_donor_haplotypes}, default is TRUE, this parameter is used to determine the threshold values used in determining whether two windows originate from the same donor. If TRUE, the preset thresholds of 0.1 and 0.9 are used.}

\item{stitch_new_min}{a numeric >0, but <1; default is 0.5; used in \code{stitch_haplotypes} within \code{impute_donor_haplotypes}, this parameter is only evaluated if \code{stringent_stitch} is FALSE and is dually assigned as the \code{different_max} and \code{same_min} threshold values when considering the concordance between two windows and therefore which donors they originate from (same or different).}

\item{smooth_imputed_genotypes}{a bool; default is FALSE; used in \code{report_gametes} whether to use smoothed data from the HMM or original reads for ending filled gamete genotypes. If TRUE, doesn't replace smoothed data from HMM with original reads when there's a mismatch}

\item{smooth_crossovers}{a bool; default is TRUE; used in \code{report_gametes} whether to use smoothed data from the HMM or original reads for recombination finding. If TRUE, doesn't replace smoothed data from HMM with original reads when there's a mismatch}
}
\value{
rhapsodi_out a named list with \code{donor_haps}, \code{gamete_haps}, \code{gamete_genotypes}, \code{unsmoothed_gamete_haps},  \code{unsmoothed_gamete_genotypes}, and \code{recomb_breaks}
}
\description{
This function runs all steps of rhapsodi by first inputting the data
Then calling \code{phase_donor_haplotypes} to run donor phasing
Then calling \code{impute_gamete_genotypes} to run gamete genotype imputation through the HMM
And finally calling \code{discover_meiotic_recombination} to run meiotic recombination discovery
The named list which is returned has
\code{donor_haps} which is the phased haplotypes as a data frame with column names index, pos (for SNP positions), h1 (haplotype 1), & h2 (haplotype 2)
\code{gamete_haps} which is the filled gametes specifying from which donor haplotype each gamete position originates in data frame form
\code{gamete_genotypes} which is the filled gametes specifying the genotype in (0's and 1's) for each gamete position as a data frame
\code{recomb_breaks}which is a data frame specifying the recombination breakpoints for each gamete
}
