% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_add_na_flatten.R
\name{sim_add_na_flatten}
\alias{sim_add_na_flatten}
\title{This function changes random indices to NA, in effect removing data}
\usage{
sim_add_na_flatten(to_change, num_nas, num_gametes, num_snps)
}
\arguments{
\item{to_change}{a matrix, the original data, with nrow of \code{num_snps} and ncol of \code{num_gametes}}

\item{num_nas}{an integer, the number of positions that should be randomly chosen to switch to NA}

\item{num_gametes}{an integer; the number of gametes or the number of columns for \code{to_return}}

\item{num_snps}{an integer; the number of SNPs or the number of rows for \code{to_return}}
}
\value{
to_return a matrix with nrow of \code{num_snps} and ncol of \code{num_gametes}. The data is a combination of locations carried over from the input \code{to_change} and locations that were switched to NAs
}
\description{
This function takes as input a matrix and carries over its data to a vector
Then \code{num_nas} number of locations are randomly chosen in that vector and overwritten by NA
Finally, the vector is reshaped to a matrix matching the size and dimension of the input and returned
This function is efficient for lower numbers of NAs (i.e. higher coverage samples)
}
