% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_add_seq_error.R
\name{sim_add_seq_error}
\alias{sim_add_seq_error}
\title{This function inverts a bit at random indices in effect adding sequencing error to data}
\usage{
sim_add_seq_error(num_snps, num_gametes, seqError_add, gam_mat_with_na)
}
\arguments{
\item{num_snps}{an integer; the number of snps or the number of rows of the input and output}

\item{num_gametes}{an integer; the number of gametes or the number of columns of the input and output}

\item{seqError_add}{a numeric; the sequencing error rate}

\item{gam_mat_with_na}{a matrix; the input sparsified gamete data with nrow of \code{num_snps} and ncol of \code{num_gametes}}
}
\value{
gam_mat_with_na a matrix; reflecting the input with \code{num_bits_to_flip} inverted bits with nrow of \code{num_snps} and ncol of \code{num_gametes}
}
\description{
This function takes as input a matrix and carries over its inverted data to a vector and its original data to a second vector
After subsetting for just non-NA locations, the number of available genotypes to change is multiplied by the sequencing error rate to find the number of bits to flip
Then these number of random (non NA) indices are sampled
and the inverted bit data is used to replace the original input data in these locations
Finallly, this data with the introduced sequencing errors is reshaped to a matrix of equal dimension and size as the input and returned
}
