% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_assess_gam_imputation.R
\name{sim_assess_gam_imputation}
\alias{sim_assess_gam_imputation}
\title{This function assesses the effectiveness of rhapsodi's gamete imputation}
\usage{
sim_assess_gam_imputation(true_gam, pred_gam, num_snps, num_gametes)
}
\arguments{
\item{true_gam}{a matrix, from the output of the generative model, the true/full gamete genotypes where the rows are the SNPs and the columns are the gametes (except for the first column which is the SNP genomic positions)}

\item{pred_gam}{a matrix, from the output of rhapsodi, with the predicted/filled gamete genotypes where the rows are the SNPs and the columns are the gametes}

\item{num_snps}{the number of snps, should be equal to the number of rows of \code{true_gam} or \code{pred_gam}}

\item{num_gametes}{the number of gametes, should be equal to the number of cols - 1 of \code{true_gam} or just the number of cols of \code{pred_gam}}
}
\value{
assess_gam a named list with vectors for \code{acc}, \code{com}, \code{lhs}, & \code{ser}
}
\description{
This function compares the simulated full truth gamete data to the rhapsodi predicted filled gamete data,
producing a named list with vectors for lhs (largest haplotype segment), ser (switch error rate), acc (accuracy), com (completeness)
}
