% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_assess_it.R
\name{sim_assess_it}
\alias{sim_assess_it}
\title{This function drives the assessment of all 3 parts of rhapsodi: phasing, gamete imputation, recombination discovery}
\usage{
sim_assess_it(
  true_donor_haps,
  pred_donor_haps,
  true_recomb,
  pred_recomb,
  true_gam,
  pred_gam,
  cons = FALSE
)
}
\arguments{
\item{true_donor_haps}{a data frame of phased donor haplotypeps from the generative model with column names of \code{donor1} and \code{donor2}}

\item{pred_donor_haps}{a tibble of phased donor haplotypes from rhapsodi with column names \code{index}, \code{pos} (for SNP positions) \code{h1} (haplotype 1), & \code{h2} (haplotype 2)}

\item{true_recomb}{a data.table data table containing the true recombination breakpoints from the generative model with columns \code{gam}, \code{start}, \code{end}}

\item{pred_recomb}{a tibble containing the predicted recombination breakpoints from rhapsodi with columns \code{Ident}, \code{Genomic_start}, \code{Genomic_end}}

\item{true_gam}{a matrix, from the output of the generative model, the true/full gamete genotypes where the rows are the SNPs and the columns are the gametes (except for the first column which is the SNP genomic positions)}

\item{pred_gam}{a matrix, from the output of rhapsodi, the predicted/filled gamete genotypes where the rows are the SNPs and the columns are the gametes}

\item{cons}{a bool; default=FALSE, If TRUE, compares recombination breakpoints in a conservative manner such that if two or more true breakpoints intersect a single predicted breakpoint, we only consider one intersection to be a tp and the rest to be fn.}
}
\value{
all_metrics a named list of named lists with all the assessment metric values or vectors
}
\description{
This function drives all 3 steps of assessment in comparing the simulated full truth data with the rhapsodi predicted data
First it assesses donor haplotypte phasing, producing a named list with single values for lhs (largest haplotype segment), ser (switch error rate), acc (accuracy), com (completeness)
Then it assesses gamete genotype imputation, producing a named list with vectors for lhs (largest haplotype segment), ser (switch error rate), acc (accuracy), com (completeness)
Then it assesses recombination discovery producing a named list with single values for precision, recall, accuracy, specificity, fdr (false discovery rate), fpr (false positive rate) f1 (f1 score),
true_n (number of true recombination breakpoints), pred_n (number of predicted recombination breakpoints), tn (true negative), fn (false negative), tp (true positive), fp (false positive)
Finally, it returns a list of named lists where \code{phasing} contains the phasing assessment named list
\code{gam_imputation} contains the gamete genotype imputation assessment named list
and \code{recomb} contains the recombination breakpoint discovery assessment named list
}
