% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_assess_phasing.R
\name{sim_assess_phasing}
\alias{sim_assess_phasing}
\title{This function assesses the effectiveness of rhapsodi's diploid donor haplotype phasing}
\usage{
sim_assess_phasing(true_donor_haps, pred_donor_haps, num_snps)
}
\arguments{
\item{true_donor_haps}{a data frame of phased donor haplotypes from the generative model with column names of \code{donor1} and \code{donor2}}

\item{pred_donor_haps}{a tibble of phased donor haplotypes from rhapsodi with column names \code{index}, \code{pos} (for SNP positions), \code{h1} (haplotype 1) & \code{h2} (haplotype 2)}

\item{num_snps}{the number of SNPs, should be equal to the number of rows of \code{true_donor_haps} or \code{pred_donor_haps}}
}
\value{
assess_phasing a named list with values for \code{acc}, \code{com}, \code{lhs}, & \code{ser}
}
\description{
This function compares the simulated truth diploid donor haplotypes to the rhapsodi predicted phased data,
producing a named list of values for lhs (largest haplotype segment), ser (switch error rate), acc (accuracy), com (completeness)
}
