% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_assess_recomb.R
\name{sim_assess_recomb}
\alias{sim_assess_recomb}
\title{This function assesses the effectiveness of rhapsodi's meiotic recombination discovery}
\usage{
sim_assess_recomb(true_recomb, pred_recomb, cons = FALSE)
}
\arguments{
\item{true_recomb}{a data.table data table containing the true recombination breakpoints from the generative model with columns \code{gam}, \code{start}, \code{end}}

\item{pred_recomb}{a tibble containing the predicted recombination breakpoints from rhapsodi with columns \code{Ident} \code{Genomic_start}, \code{Genomic_end}}

\item{cons}{a bool; default = FALSE, If TRUE, compares recombination breakpoints in a conservative manner such that if two or more true breakpoints intersect a single predicted breakpoint, we only consider one intersection to be a tp and the rest to be fn. If FALSE, all are tp.}
}
\value{
metrics_out a named list with values for \code{precision}, \code{recall}, \code{accuracy}, \code{specificity}, \code{fdr}, \code{fpr}, \code{true_n},\code{pred_n}, \code{tn}, \code{fn}, \code{tp}, \code{fp}
}
\description{
This function compares the simulated truth gamete specific recombination breakpoints to the rhapsodi predicted gamete specific breakpoints,
producing a named list with values for \code{precision}, \code{recall}, \code{accuracy}, \code{specificity}, \code{fdr} (false discovery rate), \code{fpr} (false positive rate), \code{f1} (f1 score),
\code{true_n} (number of true recombination breakpoints), \code{pred_n} (number of predicted recombination breakpoints), \code{tn} (# of true negatives), \code{fn} (# of false negatives), \code{tp} (# of true positives), \code{fp} (# of false positives)
}
