% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_completeness.R
\name{sim_completeness}
\alias{sim_completeness}
\title{This function computes the completeness of a haplotype}
\usage{
sim_completeness(predicted, num_snps)
}
\arguments{
\item{predicted}{the rhapsodi predicted haplotype vector or haplotypes matrix/data frame where genotypes are 0/1 encoded}

\item{num_snps}{an integer, the number of SNPs or the length of the haplotype}
}
\value{
to_return the completeness of the input. If \code{predicted} is a vector, \code{to_return} is a numeric. else if \code{predicted} is 2-dimensional, \code{to_return} is a vector of numerics
}
\description{
This function computes the completeness of a haplotype by finding the number of NAs
and then dividing the number of NAs by the total number of genotypes or SNPs in the haplotype
Finally, we subtract this ratio from 1 to find the ratio of complete or non-NA genotypes
in the hapltoype. Note that this function works for a single vector/haplotype or for a matrix with
multiple haplotypes. In the first case, we return a single value.
In the latter case, we return a vector of values.
}
