% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_filter_generated_data.R
\name{sim_filter_generated_data}
\alias{sim_filter_generated_data}
\title{This function filters the generated gamete data so that the sparse input data only reflects heterozygous SNPs}
\usage{
sim_filter_generated_data(
  gam_na_df,
  gam_full_df,
  donor_haps,
  new_rows,
  add_de_novo_mut
)
}
\arguments{
\item{gam_na_df}{the generated, sparsified gamete data with SNP indices in the first column}

\item{gam_full_df}{the generated full gamete data with SNP indices in the first column}

\item{donor_haps}{the generated diploid donor phased haplotypes}

\item{new_rows}{a vector of the integer SNP indices for where de novo mutation were added if DNMs were simulated, this should be NULL otherwise}

\item{add_de_novo_mt}{a bool, if TRUE de novo mutations were simulated in the generative model and then \code{new_rows} is also filtered}
}
\value{
out a named list with the filtered \code{gam_na_df}, filtered \code{gam_full_df}, filtered \code{donor_haps}, new number of snps after filtering \code{num_snps}
}
\description{
This function is used to filter the generated data such that the sparse gamete data only reflects heterozygous SNPs (one 0 and one 1)
Then we only keep these SNP indices for the full gamete data and the diploid donor haplotypes
Further, if de novo mutations were simulated, we filter the list of DNM locations to reflect those that weren't filtered out.
However, we don't return the filtered dnm list, we just output it as a message
}
