% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_find_cov_from_mgr.R
\name{sim_find_cov_from_mgr}
\alias{sim_find_cov_from_mgr}
\title{This function computes the coverage given the desired missing genotype rate}
\usage{
sim_find_cov_from_mgr(missing_genotype_rate)
}
\arguments{
\item{missing_genotype_rate}{a numeric, if greater than 1, assumed to be a percentage and is divided by 100.}
}
\value{
coverage a numeric, the coverage of the simulation corresponding to the desired missing_genotype_rate
}
\description{
This function computes the coverage of the simulation given the desired missing genotype rate
by reversing the poisson density function and taking the negative natural log of the input missing genotype rate
}
