% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_find_fp.R
\name{sim_find_fp}
\alias{sim_find_fp}
\title{This function finds the number of false positives when comparing two sets of recombination breakpoints}
\usage{
sim_find_fp(
  pred_intersect_dt,
  pred_dt_nona,
  no_truths = FALSE,
  no_preds = FALSE
)
}
\arguments{
\item{pred_intersect_dt}{a data table from foverlaps intersecting the predictions from rhapsodi and the truth data with columns of \code{gam}, \code{True_Start}, \code{True_End}, \code{Predicted_Start}, \code{Predicted_End}}

\item{pred_dt_nona}{a data table of just the predicted data from rhapsodi with no NAs in \code{start} or \code{stop} columns}

\item{no_truths}{bool, default is FALSE; if TRUE, that means there were only NAs in the truth information and no real recombination break points}

\item{no_preds}{bool, default is FALSE; if TRUE, that means there were only NAs in the prediction information and no predicted recombination break points}
}
\value{
fp an integer, the number of false negatives
}
\description{
This function finds the number of true negatives by finding the predictions which don't intersect with true recombination spots at all
If there are no non-NA truths but some non-NA predictions, the number of false positives are equal to the number of non-NA predictions
because predictions are present for breakpoints that aren't there
If there are some non-NA truths but no non-NA predictions, we have 0 false positives
}
