% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_find_mgr_from_cov.R
\name{sim_find_mgr_from_cov}
\alias{sim_find_mgr_from_cov}
\title{This function computes the missing genotype rate given the coverage}
\usage{
sim_find_mgr_from_cov(coverage)
}
\arguments{
\item{coverage}{a numeric, must be greater than 0. The desired coverage of sequencing (e.g. 0.01x, 0.1x, 1.204x, etc)}
}
\value{
missing_genotype_rate a numeric, the corresponding missing genotype rate for the given coverage, < 1
}
\description{
This function computes the missing genotype rate of the simulation given the desired coverage
by using the poisson density function parameterized by x = 0 and lambda equal to the coverage metric
to return the probability of the density function, or the missing genotype rate
Note that the function checks to assure that the coverage is greater than 0
}
