% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_find_num_crossovers.R
\name{sim_find_num_crossovers}
\alias{sim_find_num_crossovers}
\title{This function finds the number of crossovers for each gamete}
\usage{
sim_find_num_crossovers(num_gametes, recomb_lambda)
}
\arguments{
\item{num_gametes}{an integer, the number of gametes for which to generate a number of crossovers}

\item{recomb_lambda}{a numeric, the average recombination rate, used as the mean for the poisson distribution}
}
\value{
num_recomb_sites a vector of length \code{num_gametes} containing integers for the number of crossovers for each gamete
}
\description{
This function uses a poisson distribution with mean equal to \code{recomb_lambda} to randomly generate the number of crossovers
for each gamete for \code{num_gametes} gametes
}
