% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_find_num_nas.R
\name{sim_find_num_nas}
\alias{sim_find_num_nas}
\title{This function computes the number of NAs that should be used to replace genotypes given the size of the genotype matrix and the missing genotype rate}
\usage{
sim_find_num_nas(num_gametes, num_snps, missing_genotype_rate)
}
\arguments{
\item{num_gametes}{an integer, the number of gametes or the number of columns of the genotype matrix}

\item{num_snps}{an integer, the number of snps or the number of rows of the genotype matrix}

\item{missing_genotype_rate}{a numeric, < 1, the missing genotype rate}
}
\value{
num_nas an integer, the product of the missing genotype rate and the number of genotypes (\code{num_snps} * \code{num_gametes})
}
\description{
This function computes the number of NAs that are needed to achieve a certain missing genotype rate
given the size of the input genotype matrix which is num_snps rows and num_gametes columns.
}
