% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_find_tn.R
\name{sim_find_tn}
\alias{sim_find_tn}
\title{This function finds the number of true negatives when comparing two sets of recombination breakpoints}
\usage{
sim_find_tn(truth_dt_na, pred_dt_na)
}
\arguments{
\item{truth_dt_na}{a data table for the truth or generated data in which the \code{start} and \code{stop} columns have NA, meaning no recombination breakpoints}

\item{pred_dt_na}{a data table for the predicted data from rhapsodi in which the \code{start} and \code{stop} columns have NA, meaning no predicted recombination breakpoints}
}
\value{
tn an integer, the number of true negatives
}
\description{
This function find sthe number of true negatives by finding the gametes for which both truth and prediction say there are no
recombination breakpoints (NA in start and stop columns)
}
