% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_generate_global.R
\name{sim_generate_global}
\alias{sim_generate_global}
\title{This function simulates fully known gamete genotypes given a specified number of meiotic recombination events}
\usage{
sim_generate_global(num_gametes, num_snps, n_crossovers, random_seed = 42)
}
\arguments{
\item{num_gametes}{an integer, the number of gametes to simulate; also the number of columns -1 of the output}

\item{num_snps}{an integer, the number of SNPs to simulate; also the number of rows of the output}

\item{n_crossovers}{a vector of integers, whose length is \code{num_gametes}}

\item{random_seed}{an integer, the random seed which will be set for reproducibility; default is 42}
}
\value{
gam_full_df a dataframe of 0's and 1's with num_gametes + 1 columns, and num_snps rows. The first column is the SNP index and has a column name of "positions". The rest of the columns have a name corresponding to which gamete number it is (i.e. "gam101_")
}
\description{
This function runs portions of the generative model to simulate fully known gamete genotypes
for a specified number of gametes and number of SNPs
as well as a specified number of meiotic recombination events for each gamete
}
