% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{sim_metrics}
\alias{sim_metrics}
\title{This function computes and returns metrics given the number of true positives, false positives, true negatives, and false negatives that were observed in a prediction task}
\usage{
sim_metrics(tp, fp, tn, fn)
}
\arguments{
\item{tp}{an integer for the number of true positives}

\item{fp}{an integer for the number of false positives}

\item{tn}{an integer for the number of true negatives}

\item{fn}{an integer for the number of false negatives}
}
\value{
metric_list a named list returning precision, recall, accuracy, specificity, fpr, fdr, and f1 (the f1 score)
}
\description{
This function computes metrics for precision, recall, accuracy, specificity, false discovery rate, false positive rate, and the f1 score
returning the metrics in a named list and computing them from inputs of the number of true positives \code{tp},
the number of false positives \code{fp}, the number of true negatives \code{tn}, the number of false negatives \code{fn}.
For any metric, if the denominator is equal to 0, the returned value for that metric is NA.
}
