% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_with_overlap.R
\name{split_with_overlap}
\alias{split_with_overlap}
\title{A function to separate the gamete SNP positions into overlapping segments}
\usage{
split_with_overlap(vector_of_positions, window_size, overlap_denom)
}
\arguments{
\item{vector_of_positions}{vector of SNP positions}

\item{window_size}{User-input value for length of segment (default = 3000)}

\item{overlap_denom}{User-input value for denominator in calculation of overlap, or the degree of overlap between segments (default = 2)}
}
\value{
windows list of lists with the number of lists equal to the number of windows and each inner list containing the SNP positions in that overlapping segment
}
\description{
This function separates each gamete into overlapping windows of size determined by the user. These
overlaps are an input to the distance matrix constructed in \code{reconstruct_haplotypes}.
}
\details{
Adapated from https://stackoverflow.com/questions/8872376/split-vector-with-overlapping-samples-in-r
}
