% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsmooth.R
\name{unsmooth}
\alias{unsmooth}
\title{A function to unsmooth or replace original reads for the gametes}
\usage{
unsmooth(original_gamete_df, filled_gamete_data)
}
\arguments{
\item{original_gamete_df}{original gamete data with haplotype by position}

\item{filled_gamete_data}{filled gamete data from \code{fill_gametes}}
}
\value{
filled_gamete_data filled gamete data in tibble form with replaced original reads
}
\description{
This function finds where the resulting haplotype assignments differ from the original reads and replaces the imputed data with
the originally observed data, hence unsmoothing the HMM signal.
}
