% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hmm.R
\name{run_hmm}
\alias{run_hmm}
\title{A function that assigns the position in each gamete to the correct donor haplotype}
\usage{
run_hmm(dt, column_index, hmm)
}
\arguments{
\item{dt}{Matrix of gametes with 0 and 1}

\item{column_index}{\code{apply} function cycles through gamete_dt to act on each column (i.e., each gamete)}

\item{hmm}{the HMM that viterbi will be applied on}
}
\value{
original_obs Replaced observed haplotype with that assigned by the model
}
\description{
This function walks along each gamete and replaces the original observation with the inferred state.
This replaces any incorrect haplotype assignments (e.g., due to sequencing error).
}
